Rename RegistrySerialize and RegistryDeserialize to Serialize and Deserialize.
