+++ src/registry/serde/mod.rs
@@ -1,0 +1,5 @@
+mod de;
+mod ser;
+
+pub use de::Deserialize;
+pub use ser::Serialize;
